+++ __test__/tools.ts
@@ -17,3 +17,18 @@ export function getNote(id: string) {
   }
   return data;
 }
+
+export function getColoringTestObject() {
+  return {
+    todo: {
+      done_nodue: "1;2",
+      open_nodue: "3;4",
+      open: "5;6",
+      open_overdue: "7;8",
+      done: "9;10",
+      done_overdue: "11;12",
+      warning: "13;14",
+      warningHours: 15,
+    },
+  };
+}